services/data-simulator/requirements.txt
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0